Turn off HL7800 echo
Re-work the driver so it does not require echo for command processing.
This fixes issues with unsolicited notifications interrupting echoed commands.
Maximum supported sockets by HL7800 is 6.
offload_sendto: Caller will handle freeing packet if error is encountered.
Assert is checking the array size of pinconfig.  Not the actual size of the structure.

bug 14543