drivers: modem: hl800: re-init IP connection upon wakeup
When waking up from sleep (hibernate or lite hibernate, PSM or eDRX) the GPRS connection and sockets need to be restored because the networking stack is not aware that any configuration was lost.

Remove IO sleep state because it is not needed.

Set DTR as standard output pin so it can drive to logic level high properly.

Reconnect TCP socket after wake up if needed.

Signed-off-by: Ryan Erickson <ryan.erickson@lairdconnect.com>