zephyr: disk: sdhc: fix sdhc spi mode support
spi mode sdhc support is not functional in zephyr on nrf boards.

1. the const designation of the sdhc_ones data structure forces
the linker to put it in flash. however the nrf spi driver uses dma
which can only use data stored in SRAM. this generates an error code.

2. the crc7_be function always returns a value 1 less than it should for
all of the transmitted spi packets. this prevents normal operation. this
works around that problem. this change was not intended to impact
anything else utilizing that function which is why the + 1 was just
added here to limit impact elsewhere in the code.

Bug 16786