zephyr: sensor: lis2dh: fix handling of negative temperature values
this change fixes the handling of negative temperature values in the
lis2dh driver. the device reports an signed 8-bit value indicating a
delta from 25 degrees C. that value gets added to 25 (the reference
temperature) to return an absolute temperature in degrees C. however
the value read from the register is being treated as a signed 16-bit
value rather than a signed 8-bit value. the resulting incorrect
interpretation of the negative signed value results in an incorrect
value for anything below 25 degrees C (the reference temperature).

Bug 16870